requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1

langchain==0.2.7
langgraph==0.1.8

faiss-cpu==1.8.0

sentence-transformers==2.7.0
transformers==4.41.2
torch==2.2.2

groq==0.9.0

pymupdf==1.24.5
pdfplumber==0.11.0

python-multipart==0.0.9
pydantic==2.7.4
python-dotenv==1.0.1

numpy==1.26.4
tqdm==4.66.4

jinja2==3.1.4
